backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pillow==11.0.0
sentence-transformers==3.3.1
python-multipart==0.0.20
numpy==2.1.3

